Adds general database functions
Updates the database to support storing the ASCII data from files directly to the database. This combines the two tables involving receipts into one as that is all that is needed if the receipts are not going to be parsed directly.

Functions are added to support creating businesses, logging in to businesses, and uploading receipts.