Adds functionality to the sign up form
The owner name is removed from the database, and the sign up form is adjusted on the page to include a submit button, and a confirm password field instead of a owner's name field. The sign up form updates the database if the email is not in use and subsequently the page is reloaded.